requirements.txt
tensorflow>=2.16.0
numpy>=1.26.0
Pillow>=10.0.0
selenium>=4.0.0
webdriver-manager>=4.0.0
pillow>=10.0.0
